#43 Implemented expiry check.
